% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr_deshape.R
\name{apr_deshape}
\alias{apr_deshape}
\title{Remove the shape of an array}
\usage{
apr_deshape(x)
}
\arguments{
\item{x}{an array or vector}
}
\value{
a vector with dimensions removed
}
\description{
Remove the shape of an array
}
\examples{
x <- array(1:12, c(3, 4))
apr_deshape(x)  # Returns vector 1:12

x <- array(1:24, c(2, 3, 4))
apr_deshape(x)  # Returns vector 1:24
}
