% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr_index_of.R
\name{apr_index_of}
\alias{apr_index_of}
\title{Find index of first occurrence
Finds the index of the first occurrence of each element in the right argument within the left argument.}
\usage{
apr_index_of(x, y)
}
\arguments{
\item{x}{the haystack (left argument)}

\item{y}{the needle (right argument)}
}
\value{
An integer vector of the same shape as y, containing the index of
the first occurrence of each element in x. Elements not found return
length(x) + 1.
}
\description{
Find index of first occurrence
Finds the index of the first occurrence of each element in the right argument within the left argument.
}
\examples{
apr_index_of(c("A", "B", "C", "D", "E"), c("A", "C", "E"))  # Returns c(1, 3, 5)
apr_index_of(1:5, c(3, 7, 1))  # Returns c(3, 6, 1)
}
