% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{apr_reshape}
\alias{apr_reshape}
\title{Change the shape of an array}
\usage{
apr_reshape(x, value)
}
\arguments{
\item{x}{an array}

\item{value}{an integer vector}
}
\value{
an array
}
\description{
Change the shape of an array
}
