% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{apr_reverse}
\alias{apr_reverse}
\title{Reverse along the last axis}
\usage{
apr_reverse(x)
}
\arguments{
\item{x}{an array or vector}
}
\value{
an array with the last axis reversed
}
\description{
Reverses the elements along the last axis of an array. For a vector, this
reverses the entire vector. For a matrix, it reverses each row. For higher
dimensional arrays, it reverses along the last dimension.
}
\examples{
# Vector reverse
apr_reverse(1:5)  # Returns c(5, 4, 3, 2, 1)

# Matrix: reverse each row
m <- matrix(1:6, nrow = 2, ncol = 3)
apr_reverse(m)

# 3D array: reverse along last axis
arr <- array(1:24, c(2, 3, 4))
apr_reverse(arr)
}
