% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{apr_rotate}
\alias{apr_rotate}
\title{Rotate a vector}
\usage{
apr_rotate(x, k)
}
\arguments{
\item{x}{a vector}

\item{k}{an integer, the position to rotate}
}
\value{
a vector
}
\description{
Rotate a vector
}
\examples{
apr_rotate(1:12, 3)

}
