% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{apr_shape}
\alias{apr_shape}
\title{Retreive the shape of an array}
\usage{
apr_shape()
}
\arguments{
\item{x}{an array}
}
\value{
an integer vector
}
\description{
Retreive the shape of an array
}
\examples{
shape(Titanic)
}
