% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{apr_transpose}
\alias{apr_transpose}
\title{Transpose an array}
\usage{
apr_transpose(x)
}
\arguments{
\item{x}{an array}
}
\value{
an array with reversed dimensions
}
\description{
Reverses all dimensions of an array. For a 2D matrix, this is the standard
transpose. For higher dimensional arrays, it reverses the order of dimensions.
}
\examples{
# 2D matrix transpose
m <- matrix(1:6, nrow = 2, ncol = 3)
apr_transpose(m)  # 3x2 matrix

# 3D array transpose
arr <- array(1:24, c(2, 3, 4))
apr_transpose(arr)  # 4x3x2 array
}
