% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{apr_transpose_axes}
\alias{apr_transpose_axes}
\title{Transpose array axes}
\usage{
apr_transpose_axes(x, perm)
}
\arguments{
\item{x}{an array}

\item{perm}{an integer vector specifying the permutation of dimensions}
}
\value{
an array with permuted dimensions
}
\description{
Permutes the dimensions of an array according to the specified order.
This is the dyadic form of transpose, similar to R's aperm function.
}
\examples{
arr <- array(1:24, c(2, 3, 4))
apr_transpose_axes(arr, c(3, 1, 2))  # 4x2x3 array
apr_transpose_axes(arr, c(1, 3, 2))  # 2x4x3 array
}
