\name{NeuralActivity}
\alias{NeuralActivity}
\alias{NeuralScales}
\docType{data}
\title{
Neural Activity
}
\description{
The NeuralActivity data consists of a length 20 list (subjects) of 60x60 dissimilarity matrices. They were generated by correlating the patterns of neural activity elicited by thinking about another person experiencing 60 different mental states (e.g. happiness, anger, etc.). The row/column names indicate the mental states. The NeuralScales contain additional covariates for the mental states.
}
\usage{
data("NeuralActivity")
data("NeuralScales")
}
\format{
  A list of 20 dissimilarity matrices
}


\examples{
data(NeuralActivity)
str(NeuralActivity)
}
\keyword{datasets}
