% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{apr_reshape}
\alias{apr_reshape}
\title{Change the shape of an array}
\usage{
apr_reshape(x, value)
}
\arguments{
\item{x}{an array or vector}

\item{value}{an integer vector representing the new dimensions}
}
\value{
an array with the specified dimensions
}
\description{
Change the shape of an array
}
\examples{
x <- 1:12
apr_reshape(x, c(3, 4))  # Returns a 3x4 matrix
apr_reshape(x, c(2, 2, 3))  # Returns a 2x2x3 array
}
