% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{apr_shape}
\alias{apr_shape}
\title{Retrieve the shape of an array}
\usage{
apr_shape(x)
}
\arguments{
\item{x}{an array or object}
}
\value{
an integer vector representing the dimensions
}
\description{
Retrieve the shape of an array
}
\examples{
arr <- array(1:12, c(3, 2, 2))
apr_shape(arr)  # Returns c(3, 2, 2)
}
